rename: index.tsx -> Modal.tsx로 변경 - modal 폴더 안에 두 컴포넌트가 있어서 index로 하기는 애매해보임
